Move DirectoryFinder to Helpers namespace
